requirements.txt
numpy
pandas
matplotlib
seaborn
beautifulsoup4
requests
pyyaml
openpyxl
nltk
syllables
